cleaned up app.js file
